+++ layers/GatoGraphQLForWP/plugins/gatographql/src/ModuleResolvers/SchemaTypeModuleResolver.php
@@ -52,6 +52,7 @@ class SchemaTypeModuleResolver extends AbstractModuleResolver
     public final const SCHEMA_MEDIA = Plugin::NAMESPACE . '\schema-media';
     public final const SCHEMA_SITE = Plugin::NAMESPACE . '\schema-site';
     public final const SCHEMA_MULTISITE = Plugin::NAMESPACE . '\schema-multisite';
+    public final const SCHEMA_PAGEBUILDER = Plugin::NAMESPACE . '\schema-pagebuilder';
     public final const SCHEMA_TAGS = Plugin::NAMESPACE . '\schema-tags';
     public final const SCHEMA_POST_TAGS = Plugin::NAMESPACE . '\schema-post-tags';
     public final const SCHEMA_CATEGORIES = Plugin::NAMESPACE . '\schema-categories';
@@ -306,6 +307,7 @@ class SchemaTypeModuleResolver extends AbstractModuleResolver
             self::SCHEMA_COMMENTS,
             self::SCHEMA_SITE,
             self::SCHEMA_MULTISITE,
+            self::SCHEMA_PAGEBUILDER,
             self::SCHEMA_TAGS,
             self::SCHEMA_POST_TAGS,
             self::SCHEMA_CATEGORIES,
@@ -375,6 +377,7 @@ class SchemaTypeModuleResolver extends AbstractModuleResolver
             self::SCHEMA_MEDIA => \__('Media', 'gatographql'),
             self::SCHEMA_SITE => \__('Site', 'gatographql'),
             self::SCHEMA_MULTISITE => \__('Multisite', 'gatographql'),
+            self::SCHEMA_PAGEBUILDER => \__('Page Builder', 'gatographql'),
             self::SCHEMA_TAGS => \__('Tags', 'gatographql'),
             self::SCHEMA_POST_TAGS => \__('Post Tags', 'gatographql'),
             self::SCHEMA_CATEGORIES => \__('Categories', 'gatographql'),
@@ -444,6 +447,7 @@ class SchemaTypeModuleResolver extends AbstractModuleResolver
             self::SCHEMA_CUSTOMPOSTS => \__('Base functionality for all custom posts', 'gatographql'),
             self::SCHEMA_SITE => \__('Fetch site information', 'gatographql'),
             self::SCHEMA_MULTISITE => \__('Fetch site information in a WordPress multisite network', 'gatographql'),
+            self::SCHEMA_PAGEBUILDER => \__('Fetch data from Page Builders installed on the site', 'gatographql'),
             self::SCHEMA_TAGS => \__('Base functionality for all tags', 'gatographql'),
             self::SCHEMA_CATEGORIES => \__('Base functionality for all categories', 'gatographql'),
             default => parent::getDescription($module),
@@ -495,6 +499,7 @@ class SchemaTypeModuleResolver extends AbstractModuleResolver
             case self::SCHEMA_MEDIA:
             case self::SCHEMA_SITE:
             case self::SCHEMA_MULTISITE:
+            case self::SCHEMA_PAGEBUILDER:
                 return false;
         }
         return $this->upstreamHasDocumentation($module);