+++ layers/GatoGraphQLForWP/plugins/gatographql/src/PluginInitializationConfiguration.php
@@ -1093,6 +1093,9 @@ class PluginInitializationConfiguration extends AbstractMainPluginInitialization
             SchemaTypeModuleResolver::SCHEMA_MULTISITE => [
                 \PoPWPSchema\Multisite\Module::class,
             ],
+            SchemaTypeModuleResolver::SCHEMA_PAGEBUILDER => [
+                \PoPWPSchema\PageBuilder\Module::class,
+            ],
             SchemaTypeModuleResolver::SCHEMA_TAGS => [
                 TagsModule::class,
                 \PoPCMSSchema\TagsWP\Module::class,